demo file added

+++ problem2/Point.java
@@ -17,15 +17,15 @@ public class Point {
     public float  getY(){
         return this.y;
     }
-      public void setX(int x){
+      public void setX(float  x){
         this.x = x;
     }
 
-    public void setY(int y){
+    public void setY(float  y){
         this.y = y;
     }
 
-    public void setXY(int x, int y){
+    public void setXY(float  x, float  y){
         this.x = x;
         this.y = y;
     }
